ok for fucks sake i think im making a decision
migrating to Ninja build system, moving away from pure rust. this is
because a lot of the information on developing an operating system in
rust is built on top of other crates, and I want a more in depth look at
writing to msrs and other stuff like that. also llvm doesnt like when
you try to write to ebx so that shit pisses me off